rdma: fixup compile issue

+++ engines/rdma.c
@@ -46,6 +46,7 @@
 #include <inttypes.h>
 
 #include "../fio.h"
+#include "../hash.h"
 
 #ifdef FIO_HAVE_RDMA
 
@@ -1212,7 +1213,7 @@ static int fio_rdmaio_setup(struct thread_data *td)
 		rd = malloc(sizeof(*rd));;
 
 		memset(rd, 0, sizeof(*rd));
-		init_rand_seed(&rd->rand_state, GOLDEN_RATIO_PRIME);
+		init_rand_seed(&rd->rand_state, (unsigned int) GOLDEN_RATIO_PRIME);
 		td->io_ops->data = rd;
 	}
